Package file structure and license.
