--- .gitignore
@@ -1,3 +1,0 @@
-.idea/
-.ipynb_checkpoints/
-__pycache__/
