--- SubjectiveLogic/HyperopinionInterface.py
@@ -1,2 +1,0 @@
-class HyperopinionInterface:
-    pass
