--- SubjectiveLogic/__init__.py
@@ -1,4 +1,0 @@
-from SubjectiveLogic.Hyperopinion import Hyperopinion
-from SubjectiveLogic.HyperopinionInterface import HyperopinionInterface
-from SubjectiveLogic.DecisionMaking import DecisionMaking
-import SubjectiveLogic.BeliefFusion
