+++ src/subjective_logic/HyperopinionDecorator.py
@@ -1,0 +1,9 @@
+from .HyperopinionInterface import HyperopinionInterface
+
+
+class HyperopinionDecorator(HyperopinionInterface):
+    def __init__(self, hyperopinion: HyperopinionInterface) -> None:
+        self.hyperopinion: HyperopinionInterface = hyperopinion
+
+    def __getattr__(self, item):
+        return getattr(self.hyperopinion, item)
