+++ src/subjective_logic/HyperopinionInterface.py
@@ -1,0 +1,2 @@
+class HyperopinionInterface:
+    pass
