--- workspace.code-workspace
@@ -1,9 +1,0 @@
-{
-	"folders": [
-		{
-			"name": "subjective-logic-library",
-			"path": "."
-		}
-	],
-	"settings": {}
-}
